Create an abstract class to test show message action
