+++ src/test/java/org/rpgrunner/event/action/ShowMessageTest.java
@@ -1,29 +1,12 @@
 package org.rpgrunner.event.action;
 
-import junit.framework.Assert;
-import junit.framework.TestCase;
+import org.rpgrunner.GameController;
 
-import org.rpgrunner.test.helper.RandomGenerator;
-import org.rpgrunner.test.mock.GameControllerSpy;
-
-public class ShowMessageTest extends TestCase {
-    private GameControllerSpy gameController;
-    private String message;
-    private ShowMessage showMessageAction;
-
-    public void setUp() {
-        gameController = new GameControllerSpy();
-        message = RandomGenerator.getRandomString();
-        showMessageAction = new ShowMessage(gameController, message);
-    }
-
-    public void testDontExecuteBeforeCallMethod() {
-        Assert.assertNull(gameController.getLastMessage());
-    }
-
-    public void testSameMessagePassed() {
-        showMessageAction.execute();
-
-        Assert.assertSame(message, gameController.getLastMessage());
+public class ShowMessageTest extends AbstractShowMessageTest {
+    protected ShowMessage createShowMessage(
+        final GameController gameController,
+        final String message
+    ) {
+        return new ShowMessage(gameController, message);
     }
 }
